climbing: layers refactoring 2
